+++ R/get_map_data.R
@@ -48,7 +48,7 @@ get_map_data_w_cache <- memoise::memoise(function(type, params = list()) {
 get_map_data_nocache <- function(type, cache = FALSE, params = list()) {
   if (!type %in% available_sections(format = "geojson", verbose = FALSE)) {
     cli::cli_abort("You have provided type {.var {type}}
-                   which is not compatible with this function.")
+                    which is not compatible with this function.")
   }
 
   check_sf_installation(verbose = FALSE)
@@ -79,13 +79,29 @@ get_map_data_nocache <- function(type, cache = FALSE, params = list()) {
     func_params = params
   )
 
+  # Create temp directory and safe filename
+  temp_dir <- tempdir()
+  safe_filename <- tolower(type)
+  temp_file <- file.path(temp_dir, safe_filename)
+
+  writeLines(api_response, temp_file)
+
   cli::cli_progress_message("Reading data to `st`.
                               This will also take some time.")
 
+  # Clean up api response
+  rm(api_response)
+
   resp_st <- sf::st_read(
-    api_response,
+    temp_file,
     quiet = TRUE
   )
+
+  # Clean up temp file
+  rm(temp_file)
+
+  # Garbagecollect
+  gc(verbose = FALSE)
 
   cli::cli_progress_message("Converting map data to `sf` object")