+++ tests/testthat/test-1-dawa-sections.R
@@ -1,0 +1,7 @@
+test_that("All sections are applicable for dawa", {
+  for (section in available_sections()) {
+    expect_no_error(
+      dawa(section = section, per_side = 50)
+    )
+  }
+})