+++ tests/testthat/test-afstemningsomraader.R
@@ -1,4 +1,5 @@
 test_that("Voting reverse geocode returns information in different formats", {
+  skip_if_not(connection_check())
   expect_snapshot(vote_rev(x = 12.58515, y = 55.68324))
 
   expect_snapshot(vote_rev(
@@ -16,6 +17,7 @@ test_that("Voting reverse geocode returns information in different formats", {
 
 
 test_that("Afstemningsomraader autocomplete returns the right answer", {
+  skip_if_not(connection_check())
   expect_snapshot(vote_ac(input = "Dok"))
   expect_snapshot(vote_ac(input = "Æbel"))
   expect_snapshot(vote_ac(input = "Røddi"))